+++ android/app/src/main/java/com/example/activitysharing/data/domain/Event.kt
@@ -9,7 +9,7 @@ data class Event(
     val timeCreated: Long,
     val eventTime: Long,
     val createdBy: Long,
-    val numberAttending: Long,
+    val numberAttending: Int,
     val usersAttendingPreviewUrls: List<String>
 )
